+++ app/__init__.py
@@ -1,12 +1,13 @@
 """This is the entry point for the actual BCC app."""
 
+import logging
+import sys
+
+import settings
+
 # setup logger for the app
-
-import logging, os, sys
-
-level = os.getenv("LOG_LEVEL", "INFO").upper()
 logging.basicConfig(
-    level=getattr(logging, level, logging.INFO),
+    level=getattr(logging, settings.LOG_LEVEL, logging.INFO),
     format="%(asctime)s [%(levelname)-8s] %(name)s: %(message)s",
     stream=sys.stdout,
 )